Add searchtools override as part of the extension
This moves a forked version of searchtools.js into our extension. This is the
same file, minus the automatic initialization that the javascript normally
performs. Eventually, maybe this could be part of an update to the Sphinx
javascript?

Because the file is set inside `search.html` template, we can't point at our own
media server.